requirements.txt
discord.py
python-dotenv
python-asyncio
aiohttp